requirements.txt
httpx==0.27.0
schedule==1.2.1
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
google-api-python-client==2.191.0
google-auth-oauthlib==1.2.0
